Remove some duplicated logic for Certstore find_cert
Many of the implementations had the same logic of find_all_certs
followed by selecting one value arbitarily, usually first in the list,
though in the case of SQL, the last instead.

Make that behavior the default implementation.

Leave find_cert as virtual as some implementations can (Windows) or
could (MacOS, SQL) limit the number of retreived certs in order to
reduce overhead.